+++ dynamic_color/example/lib/dynamic_color_builder_example.dart
@@ -1,6 +1,5 @@
 import 'package:dynamic_color/dynamic_color.dart';
 import 'package:flutter/material.dart';
-import 'package:material_color_utilities/material_color_utilities.dart';
 
 class AdvancedExample1 extends StatelessWidget {
   const AdvancedExample1({Key? key}) : super(key: key);
@@ -10,7 +9,7 @@ class AdvancedExample1 extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return DynamicColorBuilder(
-      builder: (CorePalette? corePalette) {
+      builder: (ColorScheme? light, ColorScheme? dark) {
         return MaterialApp(
           home: Scaffold(
             body: Center(
@@ -20,13 +19,13 @@ class AdvancedExample1 extends StatelessWidget {
                   Container(
                     width: 100,
                     height: 100,
-                    // On Android S+ devices, use the 40 tone of the dynamic
-                    // primary tonal palette. Otherwise, default to a 40 tone
-                    // orange (Colors.orange.shade600).
-                    color: Color(corePalette?.primary.get(40) ?? 0xFFFB8C00),
+                    // On Android S+ devices, use the primary dynamic color.
+                    // Otherwise, default to a 40 tone orange (Colors.orange.shade600).
+                    color: light?.primary ?? const Color(0xFFFB8C00),
                   ),
                   Text(
-                      'The square\'s color is ${corePalette != null ? 'dynamic' : 'orange'}'),
+                    'The square\'s color is ${light != null ? 'dynamic' : 'orange'}',
+                  ),
                 ],
               ),
             ),